task6/requirements.txt
pypdf
psycopg2-binary
langchain
langchain-community
langchain-text-splitters
sentence-transformers
google-genai
python-dotenv